Created utils module
Moved Path functions from prior module
