+++ sqltidy/cli/dialog.py
@@ -12,6 +12,12 @@ from rich import box
 
 
 from ..rulebook import SQLTidyConfig, SUPPORTED_DIALECTS
+from ..utils import (
+    get_user_rulebooks_dir,
+    get_bundled_rulebook_path,
+    get_user_rules_dir,
+
+)
 
 
 console = Console()
@@ -87,7 +93,7 @@ def create_rulebook(
                 console.print(
                     "[dim]Proceeding with auto-generation from rules...[/dim]\n"
                 )
-                from .config_schema import generate_dialect_config
+                from ..config_schema import generate_dialect_config
 
                 config_dict = generate_dialect_config(
                     dialect, include_plugins=include_plugins
@@ -95,7 +101,7 @@ def create_rulebook(
                 config = SQLTidyConfig.from_dict(config_dict)
         else:
             # Auto-generate from rules
-            from .config_schema import generate_dialect_config
+            from ..config_schema import generate_dialect_config
 
             console.print(
                 "\n[cyan]Auto-generating config from rule metadata...[/cyan]"
@@ -277,6 +283,7 @@ def edit_rulebook(rulebook_name: Optional[str] = None) -> None:
     """
     from rich.panel import Panel
     from rich.table import Table
+    import os
 
     user_dir = get_user_rulebooks_dir()
 
@@ -585,7 +592,7 @@ def update_rulebook(
         rulebook_name: Name of the rulebook file or dialect to update, or 'all' to update all
         include_plugins: Whether to include plugin rules in the update (default: True)
     """
-    from .config_schema import generate_dialect_config
+    from ..config_schema import generate_dialect_config
     from rich.panel import Panel
     from rich.table import Table
     from rich.tree import Tree
@@ -881,17 +888,6 @@ def update_rulebook(
         console.print("\n[yellow]Update cancelled.[/yellow]")
 
 
-
-
-def get_user_rulebooks_dir() -> Path:
-    """Get the path to user's rulebook directory."""
-    return Path.home() / ".sqltidy" / "rulebooks"
-
-def get_bundled_rulebooks_dir() -> Path:
-    """Get the path to bundled rulebook templates."""
-    return Path(__file__).parent / "rulebooks"
-
-
 def initialize_user_rulebooks() -> None:
     """
     Initialize user rulebook directory.
@@ -917,7 +913,7 @@ def initialize_user_rulebooks() -> None:
             return
 
     # No bundled files - generate from rules
-    from .config_schema import save_dialect_config_to_json
+    from ..config_schema import save_dialect_config_to_json
 
     for dialect in SUPPORTED_DIALECTS:
         user_rulebook = user_dir / f"sqltidy_{dialect}.json"
@@ -926,39 +922,6 @@ def initialize_user_rulebooks() -> None:
                 dialect, str(user_rulebook), include_plugins=False
             )
 
-
-def get_rulebook_path(dialect: str) -> Path:
-    """
-    Get the path to a rulebook file, checking user directory first, then bundled.
-
-    Args:
-        dialect: SQL dialect name
-
-    Returns:
-        Path to the rulebook file (user rulebook if exists, otherwise bundled)
-    """
-    user_path = get_user_rulebooks_dir() / f"sqltidy_{dialect}.json"
-    if user_path.exists():
-        return user_path
-    return get_bundled_rulebook_path(dialect)
-
-
-def get_bundled_rulebook_path(dialect: str) -> Path:
-    """
-    Get the path to a bundled rulebook template for a specific dialect.
-
-    Args:
-        dialect: SQL dialect name
-
-    Returns:
-        Path to the bundled rulebook file
-    """
-    return get_bundled_rulebooks_dir() / f"sqltidy_{dialect}.json"
-
-
-def get_default_filename(dialect: str) -> str:
-    """Get default rulebook filename for a dialect."""
-    return f"sqltidy_{dialect}.json"
 
 
 
@@ -978,10 +941,6 @@ def load_rulebook_file(filepath: str) -> Dict[str, Any]:
 
 # rule Management
 
-
-def get_user_rules_dir() -> Path:
-    """Get the path to user's rule directory."""
-    return Path.home() / ".sqltidy" / "rules"
 
 
 def add_rule(rule_file: str) -> None:
